Safari-611.1.10/LayoutTests/fast/animation/css-animation-resuming-when-visible-expected.txt

Tests that CSS animations that are created while the page is hidden are properly resumed when the page becomes visible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.animationsAreSuspended() is true
PASS internals.numberOfActiveAnimations() is 0
PASS internals.numberOfActiveAnimations() became 1
PASS successfullyParsed is true

TEST COMPLETE
TEST
